all: remove ExprKindChance
It doesn't seem particularly useful and we still need to have a bunch
of workarounds (e.g. even if only Unary is enabled we still generate
Binary for strings, and even if only FuncCalls are enabled we still
generate Binary if a Func with the required type is not in scope).

Also fix a bug in the driver in the BuildDeps code.